+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html) with
 the exception that the versions 0.*.* may have breaking changes in minor versions.
 
+## [0.5.4]
+### Removed
+- Removed `Account` class `short_qname` attribute. The `to_polars` and
+  `write_txns` methods of `Journal` class now uses a `short_qname_lenght`
+  parameter to limit the shortest qualified name of an account.
+
 ## [0.5.3]
 ### Fixed
 - Fixed importation of empty strings as tags in `from_csv` method of `Journal`